Updated review value include halfs

+++ PassedPawn.Models/DTOs/Course/CoachCourseDetailsDto.cs
@@ -8,7 +8,7 @@ public class CoachCourseDetailsDto
     public required string Title { get; init; }
     public required string Description { get; init; }
     public CourseStatus Status { get; init; }
-    public double AverageScore { get; init; }
+    public decimal AverageScore { get; init; }
     public float Price { get; init; }
     public int EnrolledStudentsCount { get; init; }
     public string? ThumbnailUrl { get; init; }
