+++ PassedPawn.Models/DTOs/Course/CoachCourseDto.cs
@@ -8,7 +8,7 @@ public class CoachCourseDto
     public required string Title { get; init; }
     public required string Description { get; init; }
     public CourseStatus Status { get; init; }
-    public double AverageScore { get; init; }
+    public decimal AverageScore { get; init; }
     public float Price { get; init; }
     public int? EloRangeStart { get; init; }
     public int? EloRangeEnd { get; init; }
