+++ PassedPawn.Models/DTOs/Course/Review/CourseReviewUpsertDto.cs
@@ -4,6 +4,6 @@ namespace PassedPawn.Models.DTOs.Course.Review;
 
 public class CourseReviewUpsertDto
 {
-    [Range(1, 5)] public int Value { get; init; }
+    [Range(1, 5)] public decimal Value { get; init; }
     public string? Content { get; init; }
 }
